Refactor foundation model into its own predictor
This will help later as more tasks are folded into the common predictor
- like layout.

Currently implements a singleton pattern - Where the model and processor
are only loaded once and initialized. This is to avoid multiple model
copies when multiple predictors subclass the FoundationPredictor. Since
steps happen sequentially, this shouldn't be a problem